ColorLine rewritten to TS
