+++ src/components/ColorLine/ColorLine.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import ColorLine from '@components/ColorLine'
+import { ColorLine } from '@components/ColorLine'
 import Spacer from '@components/Spacer'
 
 # ColorLine
