+++ src/components/ColorLine/ColorLine.test.tsx
@@ -1,8 +1,7 @@
 import '@testing-library/jest-dom'
 
-import ColorLine from '@components/ColorLine'
+import { ColorLine } from '@components/ColorLine'
 import { render, screen } from '@testing-library/react'
-import React from 'react'
 
 describe('<ColorLine /> mount', () => {
   it('is mounted', () => {
