+++ src/components/index.ts
@@ -13,7 +13,7 @@ export { default as Card } from './Card'
 export { default as Check } from './Check'
 export { default as CheckboxNoFormik } from './Checkbox'
 export * from './Close'
-export { default as ColorLine } from './ColorLine'
+export * from './ColorLine'
 export { default as ColorNumber } from './ColorNumber'
 export { default as CustomLink } from './CustomLink'
 export { default as DateTimePicker } from './DateTimePicker'
